+++ examples/composition/repeater.h
@@ -24,14 +24,14 @@ class Repeater {
 
   void Clear() noexcept;
   void Pop() noexcept;
-  void Push(std::optional<float> pitch_or, int length = 1) noexcept;
+  void Push(std::optional<double> pitch_or, int length = 1) noexcept;
 
   void SetRate(double rate) noexcept;
-  void Start(float pitch_offset = 0.0) noexcept;
+  void Start(double pitch_offset = 0.0) noexcept;
   void Stop() noexcept;
 
   void SetMode(Mode mode) noexcept { mode_ = mode; }
-  void SetNoteCallback(std::function<void(float)> note_callback) noexcept {
+  void SetNoteCallback(std::function<void(double)> note_callback) noexcept {
     note_callback_ = std::move(note_callback);
   }
 
@@ -46,12 +46,12 @@ class Repeater {
   Performer performer_;
   Task task_;
 
-  std::function<void(float)> note_callback_;
-  std::vector<std::pair<std::optional<float>, int>> pitches_;
+  std::function<void(double)> note_callback_;
+  std::vector<std::pair<std::optional<double>, int>> pitches_;
 
   Mode mode_ = Mode::kForward;
   int index_ = -1;
-  float pitch_offset_ = 0.0;
+  double pitch_offset_ = 0.0;
   int remaining_length_ = 0;
 };
